locale: Fixed compiler warning on Visual Studio.

+++ src/locale/windows/SDL_syslocale.c
@@ -78,7 +78,7 @@ SDL_SYS_GetPreferredLocales_vista(char *buf, size_t buflen)
     if (!pGetUserPreferredUILanguages(MUI_LANGUAGE_NAME, &numlangs, wbuf, &wbuflen)) {
         SDL_SYS_GetPreferredLocales_winxp(buf, buflen);  /* oh well, try the fallback. */
     } else {
-        const ULONG endidx = SDL_min(buflen, wbuflen - 1);
+        const ULONG endidx = (ULONG) SDL_min(buflen, wbuflen - 1);
         ULONG str_start = 0;
         ULONG i;
         for (i = 0; i < endidx; i++) {
